[Tizen] Fix to call ewk_shutdown only when app is terminated
ewk_init and ewk_shutdown are designed to be called only when app is
initialized and app is terminated.
The singleton instance of WebEngineManager can be destructed before app
is terminated.
So it has been fixed that destructor of WebEngineManager does not call
ewk_shutdown and ewk_shutdown is only called when app is terminated.

Change-Id: I763e968f44fb18a59572bc48c714b06076cf2a78