(Vector) Let we use processor event only once, for fast rasterization case
Let we process processor only once, for take EventProcess loop as fast as we can.
But if we fail to catch that loop case,
it mean Vector rasterization was not fast enough, so we cannot assume when
it should be finished.

Previous Processor registered for the fast-finished rasterization case that
will not show some 'flickering'. But here, if at least 1 event loop progressed,
that 'flickering' case is not avoidable.

Since call Processor every event loop might give some overhead,
let we just process it only 1 times per event loop, and believe eventfd.

Change-Id: Iefb1ae75d98e41e14ae73d4a69a97747cfdad1e6
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>